requirements.txt
PySide6>=6.0.0
pycryptodome>=3.15.0
PyYAML>=6.0
paramiko>=2.8.0
packaging>=21.3
psutil>=5.9.0
requests>=2.28.0
urllib3>=1.26.0
